Add bonus fruits to gameplay maze
The fruits appear after pacman has eaten 30 % of the pellets and again after
he has eaten 70 % of the pellets. The bonus points awarded to the player after
eating a fruit increases as the game levels increases. In addition, the bonus
fruit appears for a predetermined amount of time before it disappears if left
uneaten.